Refactor schema parsing code
The tag/attribute parser is now contained in three classes, and the
SchemaAttributes class has been moved to another class in the interim
until the new entity classes are created.
